+++ fefuactivity/Assets.xcassets/greyActivity.imageset/greyActivity.pdf
@@ -1,0 +1,260 @@
+%PDF-1.7
+
+1 0 obj
+  << /BBox [ 0.000000 0.000000 65.357590 41.956543 ]
+     /Resources << >>
+     /Subtype /Form
+     /Length 2 0 R
+     /Group << /Type /Group
+               /S /Transparency
+            >>
+     /Type /XObject
+  >>
+stream
+q
+1.000000 0.000000 -0.000000 1.000000 0.000000 35.861938 cm
+0.000000 0.000000 0.000000 scn
+3.955622 1.506919 m
+1.402258 1.506919 l
+0.828797 -0.083701 l
+0.000000 -0.083701 l
+2.327329 6.010888 l
+3.030551 6.010888 l
+5.362066 -0.083701 l
+4.537455 -0.083701 l
+3.955622 1.506919 l
+h
+1.645037 2.168283 m
+3.717029 2.168283 l
+2.678940 5.018843 l
+1.645037 2.168283 l
+h
+f
+n
+Q
+q
+1.000000 0.000000 -0.000000 1.000000 0.000000 35.694519 cm
+0.000000 0.000000 0.000000 scn
+14.773015 2.017578 m
+14.697670 1.372958 14.459077 0.874843 14.057236 0.523232 c
+13.658185 0.174412 13.126583 0.000001 12.462429 0.000001 c
+11.742464 0.000001 11.164818 0.258129 10.729489 0.774383 c
+10.296952 1.290637 10.080684 1.981301 10.080684 2.846375 c
+10.080684 3.432394 l
+10.080684 3.998878 10.181144 4.496993 10.382065 4.926740 c
+10.585775 5.356486 10.873203 5.685773 11.244349 5.914599 c
+11.615493 6.146215 12.045239 6.262024 12.533588 6.262024 c
+13.180999 6.262024 13.700044 6.080637 14.090722 5.717864 c
+14.481401 5.357882 14.708832 4.858371 14.773015 4.219332 c
+13.965147 4.219332 l
+13.895383 4.704890 13.743298 5.056501 13.508890 5.274165 c
+13.277273 5.491829 12.952172 5.600661 12.533588 5.600661 c
+12.020124 5.600661 11.616889 5.410903 11.323879 5.031386 c
+11.033661 4.651869 10.888552 4.111896 10.888552 3.411464 c
+10.888552 2.821260 l
+10.888552 2.159897 11.026685 1.633876 11.302950 1.243197 c
+11.579216 0.852519 11.965709 0.657179 12.462429 0.657179 c
+12.908919 0.657179 13.250763 0.757639 13.487961 0.958560 c
+13.727949 1.162271 13.887012 1.515277 13.965147 2.017578 c
+14.773015 2.017578 l
+h
+f
+n
+Q
+q
+1.000000 0.000000 -0.000000 1.000000 0.000000 35.861938 cm
+0.000000 0.000000 0.000000 scn
+24.116991 5.349524 m
+22.158014 5.349524 l
+22.158014 -0.083701 l
+21.358519 -0.083701 l
+21.358519 5.349524 l
+19.403730 5.349524 l
+19.403730 6.010888 l
+24.116991 6.010888 l
+24.116991 5.349524 l
+h
+f
+n
+Q
+q
+1.000000 0.000000 -0.000000 1.000000 0.000000 35.861938 cm
+0.000000 0.000000 0.000000 scn
+30.041130 -0.083701 m
+29.237450 -0.083701 l
+29.237450 6.010888 l
+30.041130 6.010888 l
+30.041130 -0.083701 l
+h
+f
+n
+Q
+q
+1.000000 0.000000 -0.000000 1.000000 0.000000 35.861938 cm
+0.000000 0.000000 0.000000 scn
+37.689838 0.983689 m
+39.431152 6.010888 l
+40.310177 6.010888 l
+38.049824 -0.083701 l
+37.338230 -0.083701 l
+35.082058 6.010888 l
+35.956902 6.010888 l
+37.689838 0.983689 l
+h
+f
+n
+Q
+q
+1.000000 0.000000 -0.000000 1.000000 0.000000 35.861938 cm
+0.000000 0.000000 0.000000 scn
+46.150608 -0.083701 m
+45.346924 -0.083701 l
+45.346924 6.010888 l
+46.150608 6.010888 l
+46.150608 -0.083701 l
+h
+f
+n
+Q
+q
+1.000000 0.000000 -0.000000 1.000000 0.000000 35.861938 cm
+0.000000 0.000000 0.000000 scn
+55.867119 5.349524 m
+53.908142 5.349524 l
+53.908142 -0.083701 l
+53.108646 -0.083701 l
+53.108646 5.349524 l
+51.153858 5.349524 l
+51.153858 6.010888 l
+55.867119 6.010888 l
+55.867119 5.349524 l
+h
+f
+n
+Q
+q
+1.000000 0.000000 -0.000000 1.000000 0.000000 35.861938 cm
+0.000000 0.000000 0.000000 scn
+62.854458 2.951036 m
+64.445076 6.010888 l
+65.357590 6.010888 l
+63.256298 2.189212 l
+63.256298 -0.083701 l
+62.452618 -0.083701 l
+62.452618 2.189212 l
+60.351322 6.010888 l
+61.272209 6.010888 l
+62.854458 2.951036 l
+h
+f
+n
+Q
+q
+1.000000 0.000000 -0.000000 1.000000 22.858559 2.909119 cm
+0.000000 0.000000 0.000000 scn
+14.488926 5.826225 m
+5.136286 5.826225 l
+3.035775 -0.000010 l
+0.000000 -0.000010 l
+8.524701 22.323669 l
+11.100511 22.323669 l
+19.640545 -0.000010 l
+16.620102 -0.000010 l
+14.488926 5.826225 l
+h
+6.025553 8.248713 m
+13.614991 8.248713 l
+9.812606 18.689939 l
+6.025553 8.248713 l
+h
+f
+n
+Q
+q
+1.000000 0.000000 -0.000000 1.000000 21.856047 0.000000 cm
+0.000000 0.000000 0.000000 scn
+10.796740 4.419203 m
+18.006119 25.232819 l
+21.645470 25.232819 l
+12.287140 -0.000008 l
+9.341000 -0.000008 l
+0.000000 25.232819 l
+3.622020 25.232819 l
+10.796740 4.419203 l
+h
+f
+n
+Q
+
+endstream
+endobj
+
+2 0 obj
+  3800
+endobj
+
+3 0 obj
+  << /XObject << /X1 1 0 R >>
+     /ExtGState << /E1 << /ca 0.500000 >> >>
+  >>
+endobj
+
+4 0 obj
+  << /Length 5 0 R >>
+stream
+/DeviceRGB CS
+/DeviceRGB cs
+q
+/E1 gs
+/X1 Do
+Q
+
+endstream
+endobj
+
+5 0 obj
+  46
+endobj
+
+6 0 obj
+  << /Annots []
+     /Type /Page
+     /MediaBox [ 0.000000 0.000000 65.357590 41.956543 ]
+     /Resources 3 0 R
+     /Contents 4 0 R
+     /Parent 7 0 R
+  >>
+endobj
+
+7 0 obj
+  << /Kids [ 6 0 R ]
+     /Count 1
+     /Type /Pages
+  >>
+endobj
+
+8 0 obj
+  << /Type /Catalog
+     /Pages 7 0 R
+  >>
+endobj
+
+xref
+0 9
+0000000000 65535 f
+0000000010 00000 n
+0000004058 00000 n
+0000004081 00000 n
+0000004177 00000 n
+0000004279 00000 n
+0000004300 00000 n
+0000004473 00000 n
+0000004547 00000 n
+trailer
+<< /ID [ (some) (id) ]
+   /Root 8 0 R
+   /Size 9
+>>
+startxref
+4606
+%%EOF
