+++ fefuactivity/Controllers/LoginController.swift
@@ -1,0 +1,16 @@
+//
+//  LoginController.swift
+//  fefuactivity
+//
+//  Created by Mike Litvin on 15.10.2021.
+//
+
+import UIKit
+
+class LoginController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+    }
+
+}
