+++ fefuactivity/RegistrationButton.swift
@@ -1,14 +1,29 @@
 //
-//  RegistrationButton.swift
+//  registrationButton.swift
 //  fefuactivity
 //
-//  Created by Mike Litvin on 07.10.2021.
+//  Created by Mike Litvin on 15.10.2021.
 //
 
 import UIKit
+@IBDesignable
 
-class RegistrationButton: UIButton {
-    func setup(){
-    layer.cornerRadius = 6;
+class registrationButton: UIButton {
+    @IBInspectable var borderWidth: CGFloat = 0.0 {
+        didSet {
+            self.layer.borderWidth = borderWidth
+        }
+    }
+    
+    @IBInspectable var borderRadius: CGFloat = 0.0 {
+        didSet {
+            self.layer.cornerRadius = borderRadius
+        }
+    }
+    
+    @IBInspectable var borderColor: UIColor = .clear {
+        didSet {
+            self.layer.borderColor = borderColor.cgColor
+        }
     }
 }
