--- fefuactivity/RegistrationController.swift
@@ -1,15 +1,0 @@
-//
-//  RegistrationController.swift
-//  fefuactivity
-//
-//  Created by Mike Litvin on 07.10.2021.
-//
-
-import UIKit
-
-class RegistrationController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-    }
-}
